requirements.txt
streamlit
pandas
numpy
plotly
python-dotenv
requests
websocket-client
scipy
neo-api-client
